Replacing T4MVC with Expression-based URL helpers.
I've decided that it's an anti-pattern to "generate" URLs in a view. Any
URL should be part of the view model (or just use magic strings
for _very simple_ URLs only).

This allows us to use "ExpressionHelper.GetRouteValuesFromExpression"
to generate URLs with strongly typed Expressions. It may not be the fastest
approach, and I have not benchmarked, but it's simple and allows for simpler
refactoring.